recommendation_engine/requirements.txt
fastapi
uvicorn
pymongo
pandas
scikit-learn
numpy
python-dotenv
motor
python-multipart
